Last ILCompiler.TypeSystem <-> ILCompiler.TypeSystem.ReadyToRun diff
After this and dotnet/runtime#63280 there will be no differences between ILCompiler.TypeSystem and ILCompiler.TypeSystem.ReadyToRun and we can unify them.
