+++ src/coreclr/tools/aot/ILCompiler.TypeSystem/ILCompiler.TypeSystem.csproj
@@ -100,9 +100,6 @@
     <Compile Include="..\..\Common\TypeSystem\Canon\TypeSystemContext.Canon.cs">
       <Link>TypeSystem\Canon\TypeSystemContext.Canon.cs</Link>
     </Compile>
-    <Compile Include="..\..\Common\TypeSystem\Aot\TargetDetails.Aot.cs">
-      <Link>TypeSystem\Aot\TargetDetails.Aot.cs</Link>
-    </Compile>
     <Compile Include="..\..\Common\TypeSystem\CodeGen\FieldDesc.CodeGen.cs">
       <Link>TypeSystem\CodeGen\FieldDesc.CodeGen.cs</Link>
     </Compile>
@@ -180,9 +177,6 @@
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Common\Utilities\GCPointerMap.Algorithm.cs">
       <Link>Utilities\GCPointerMap.Algorithm.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Common\TypeSystem\Common\Utilities\GCPointerMap.Aot.cs">
-      <Link>Utilities\GCPointerMap.Aot.cs</Link>
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Common\Utilities\GCPointerMap.cs">
       <Link>Utilities\GCPointerMap.cs</Link>
@@ -808,6 +802,5 @@
     <Compile Include="..\..\Common\TypeSystem\Sorting\TypeSystemComparer.cs">
       <Link>TypeSystem\Sorting\TypeSystemComparer.cs</Link>
     </Compile>
-    <Compile Include="Utilities\UniqueTypeNameFormatter.cs" />
   </ItemGroup>
 </Project>
